Move ServerListing to common
